Fixed a problem I caused
I changed the "dir" variable in HomePostProcess.processEpisode() to
"directory" because "dir" is a reserved built-in symbol that triggered
a warning. This method is exposed to cherrypy though, so I also needed
to change the name of the parameter in the manual post processing
template and the autoprocess script.

I also fixed a minor possible problem where a getURL call wants to handle
exceptions but I wasn't throwing them.